rushil-neura-project/python_backend/requirements.txt
Flask
Flask_Cors
flask_session
joblib
matplotlib
nltk
numpy
pandas
plotly
pymongo
python-dotenv
python_jose
redis
Requests
scikit_learn
seaborn
spacy
vaderSentiment
waitress
langchain-text-splitters
openai
numpy
tabulate
pinecone
logging
psutil
